+++ src/index.js
@@ -15,6 +15,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+const get = require('lodash.get');
 const { fn, exec, fnExport } = require('./fn');
 const { json } = require('./json');
 
@@ -28,3 +29,9 @@ module.exports.exec = exec;
 
 // name exports from json
 module.exports.json = json;
+
+// get value from the object using lodash.get
+// when object is resolved
+module.exports.get = (objOrPromise, path, options = {}) =>
+  fn(o => get(o, path), options)
+  (objOrPromise);